Change the pom to package hpi file with resources from hub-common
Use the remote-resources plugin and the resources plugin to copy the
files in the webapps folder into the shared resources folder for the
risk report.  Then use the shared resources folder for the risk report
to include the web resources folder.  Remove the changes to copy the
files from the jar file.